turboquant-hip : integrate precision fixes, engine unification, and heuristic audit remediation
- Fix SIMDS_PER_CU=2->4 (RDNA 2 has 4 SIMD32 per CU, not 2) in occupancy solver
- Fix telemetry parser to use duration-weighted averages (not last-kernel overwrite)
- Remove dead duplicate supports_turbo property in mlnn.py
- Create unified engines/run_diagnostics.py (subsumes mlnn_v40_runner.py + master_debug_turbo.py)
- Add GGML_ASSERT guards for WebGPU buffer memset >4GB truncation
- Fix MoE top-k switch: GGML_ABORT on unrecognized expert count
- Remove redundant ggml_backend_synchronize(input_backend) from async copy fallback
- Convert size_t->int truncation to proper size_t types in KleidiAI path
- Centralize float extraction from op_params in ops.cpp with typed helper
- CUDA backend fixes: VMM pool best-effort cleanup, pool eviction at >90% usage,
  null-guard CUDA destructor on teardown, P2P falls back to host-staged copy,
  hardcoded block stride replaced with sizeof()
- Grammar struct: add proper constructor, delete copy ctor/assignment to prevent dangling pointers
- Server: gate request/response body logging behind log_request_bodies flag (default OFF)

+++ ggml/src/ggml-cpu/kleidiai/kleidiai.cpp
@@ -1022,7 +1022,7 @@ class tensor_traits : public ggml::cpu::tensor_traits {
         // with cols == n.
         const size_t nchunk_size = std::max<size_t>(1, ceil_div_size(n, chunk_cols));
         GGML_ASSERT(nchunk_size <= (size_t)INT_MAX);
-        const int nchunk = (int)nchunk_size;
+        const size_t nchunk = nchunk_size;
         const size_t dst_stride = dst->nb[1];
 
         auto run_chunk = [&](runtime_slot & slot, size_t global_start, size_t cols, uint8_t * dst_batch_base) {
@@ -1095,7 +1095,7 @@ class tensor_traits : public ggml::cpu::tensor_traits {
             ggml_barrier(params->threadpool);
 
             runtime_slot & slot = runtime[local_slot];
-            int current_chunk = ith_total;
+            size_t current_chunk = ith_total;
             while (current_chunk < nchunk) {
                 const size_t global_start = (size_t)current_chunk * chunk_cols;
                 if (global_start >= n) {
